Factor out HeroTrait from ELementTrait
